+++ style/style.css
@@ -65,3 +65,8 @@
 #ranking-field{
   float : right;
 }
+
+#no-search-results{
+  font-size: 16px;
+  margin: 10% auto;
+}
